+++ src/main/java/com/telepathicgrunt/blame/mixin/MinecraftServerMixin.java
@@ -1,0 +1,55 @@
+package com.telepathicgrunt.blame.mixin;
+
+import com.mojang.authlib.GameProfileRepository;
+import com.mojang.authlib.minecraft.MinecraftSessionService;
+import com.mojang.datafixers.DataFixer;
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.DataResult;
+import com.mojang.serialization.DynamicOps;
+import com.telepathicgrunt.blame.main.DynamicRegistryManagerBlame;
+import com.telepathicgrunt.blame.main.RegistryOpsBlame;
+import net.minecraft.resource.ResourceManager;
+import net.minecraft.resource.ResourcePackManager;
+import net.minecraft.resource.ServerResourceManager;
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.WorldGenerationProgressListenerFactory;
+import net.minecraft.util.Identifier;
+import net.minecraft.util.UserCache;
+import net.minecraft.util.dynamic.RegistryOps;
+import net.minecraft.util.registry.DynamicRegistryManager;
+import net.minecraft.util.registry.Registry;
+import net.minecraft.util.registry.RegistryKey;
+import net.minecraft.util.registry.SimpleRegistry;
+import net.minecraft.world.SaveProperties;
+import net.minecraft.world.level.storage.LevelStorage;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
+
+import java.net.Proxy;
+import java.util.Collection;
+import java.util.Iterator;
+
+/* @author - TelepathicGrunt
+ *
+ * A mixin to print unregistered worldgen stuff
+ *
+ * LGPLv3
+ */
+@Mixin(value=MinecraftServer.class, priority = 10000)
+public class MinecraftServerMixin {
+
+	/**
+	 * The main hook for the parser to work from. This will check every biomes in the
+	 * DynamicRegistry to see if it has exploded due to unregistered stuff added to it.
+	 */
+	@Inject(method = "<init>",
+			at = @At(value = "RETURN"), require = 1)
+	private void init(Thread thread, DynamicRegistryManager.Impl impl, LevelStorage.Session session, SaveProperties saveProperties, ResourcePackManager resourcePackManager, Proxy proxy, DataFixer dataFixer, ServerResourceManager serverResourceManager, MinecraftSessionService minecraftSessionService, GameProfileRepository gameProfileRepository, UserCache userCache, WorldGenerationProgressListenerFactory worldGenerationProgressListenerFactory, CallbackInfo ci)
+	{
+		DynamicRegistryManagerBlame.printUnregisteredWorldgenConfiguredStuff(impl);
+	}
+}